requirements.txt
Django==3.2.13
uwsgi
django-taggit
django_widget_tweaks
django-ckeditor-5
channels_redis
psycopg2
Pillow
requests
django-debug-toolbar
django-cleanup
django-unused-media
python-dateutil
